hh_bot/requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
pyasn1==0.5.0
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.3
